Refactor package structure: remove `service` and use `services`, update imports accordingly
