+++ src/main/java/org/example/alfs/controllers/AttachmentController.java
@@ -2,7 +2,7 @@ package org.example.alfs.controllers;
 
 import org.example.alfs.dto.attachment.AttachmentViewDTO;
 import org.example.alfs.entities.Attachment;
-import org.example.alfs.service.AttachmentService;
+import org.example.alfs.services.AttachmentService;
 import org.example.alfs.repositories.AttachmentRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
